Fix #64: minor, type warnings in 64-bit time_t on 32-bit platforms
 - logit(), when formatting microsecond timestamp in stdout logging
 - scaletime(), used by mroutectl and others to dump time info

+++ src/log.c
@@ -122,8 +122,8 @@ void logit(int severity, int syserr, const char *format, ...)
 	thyme = localtime(&now_sec);
 //	if (!debug)
 	    fprintf(stderr, "%s: ", log_name);
-	fprintf(stderr, "%02d:%02d:%02d.%03ld %s", thyme->tm_hour,
-		thyme->tm_min, thyme->tm_sec, now.tv_usec / 1000, msg);
+	fprintf(stderr, "%02d:%02d:%02d.%03d %s", thyme->tm_hour,
+		thyme->tm_min, thyme->tm_sec, (int)(now.tv_usec / 1000), msg);
 	if (syserr == 0)
 	    fprintf(stderr, "\n");
 	else
